chore: Update EC2 and EKS module sources
The code changes in `_init.tf` update the source paths for the `ec2` and `eks` modules. The paths are changed to "./modules/ec2" and "./modules/eks" respectively. This update is necessary to ensure the correct module sources are used.

The recent user commits and repository commits suggest a convention of using the "chore" prefix for updates and modifications to the infrastructure code.